Generalise CLM com fix
