+++ esmvalcore/cmor/_fixes/cordex/ichec_ec_earth/clmcom_cclm4_8_17.py
@@ -1,4 +1,4 @@
 """Fixes for rcm CLMcom-CCLM4-8-17 driven by ICHEC-EC-EARTH."""
-from ..miroc_miroc5.clmcom_cclm4_8_17 import AllVars as BaseFix
+from ..cordex_fixes import CLMcomCCLM4817 as BaseFix
 
 AllVars = BaseFix
