+++ esmvalcore/cmor/_fixes/cordex/miroc_miroc5/__init__.py
@@ -1,0 +1,1 @@
+"""Fixes for CORDEX data."""
